python/python-adk-a2a/pyproject.toml
[project]
name = "python-adk-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.5",
    "fastapi>=0.115.12",
    "google-adk>=1.2.1",
    "gradio>=5.33.1",
    "mcp>=1.9.3",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
